requirements.txt
transformers>=4.36.0
datasets>=2.14.0
accelerate>=0.24.0
bitsandbytes>=0.41.0
pyyaml>=6.0.1
pytest>=7.4.0
tqdm>=4.66.0